requirements.txt
behave==1.2.6
certifi==2018.1.18
chardet==3.0.4
deepdiff==3.3.0
firebirdsql==1.1.0
idna==2.6
jsonpatch==1.21
jsonpickle==0.9.6
jsonpointer==2.0
Logbook==1.3.0
lxml==4.1.1
netifaces==0.10.6
parse==1.8.2
parse-type==0.4.2
Pillow==5.0.0
py==1.5.2
pypiwin32==223
pyshark==0.3.7.11
pywin32==223
requests==2.18.4
selenium==3.10.0
six==1.11.0
texttable==1.2.1
trollius==1.0.4
urllib3==1.22
WMI==1.4.9
zeroconf==0.20.0